microservices_narutalk/requirements.txt
# Django 관리 시스템
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# FastAPI 에이전트 서비스
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2

# HTTP 클라이언트 (마이크로서비스 간 통신)
httpx==0.25.0
aiohttp==3.8.6

# OpenAI API
openai==1.3.0

# 데이터베이스
chromadb==0.4.15

# 임베딩 모델
sentence-transformers==2.2.2
torch==2.0.1
transformers==4.35.0

# 유틸리티
python-dotenv==1.0.0
pydantic-settings==2.0.3
python-multipart==0.0.6

# 로깅 및 모니터링
structlog==23.1.0

# 동시 실행
multiprocessing-logging==0.3.4
psutil==5.9.5 